+++ src/filemerge.py
@@ -1,7 +1,5 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-from __future__ import print_function
-# ===========================================================================
 """c64tools - filemerge - Joins a set of files into one.
 
 Loads the files given and stores them at the designated
@@ -9,15 +7,22 @@ Loads the files given and stores them at the designated
 Saves the so obtained memory part, together with the
 starting address.
 
---file/-f: the file(s) to load, separated by a ','
---output/-o: the `name of the file to write
 
-(c) Daniel Krajzewicz 2016-2024
-daniel@krajzewicz.de
+Arguments
+---------
+
+* OUTPUT_FILE: Defines the name of the file to write
+* INPUT_FILE(s): Defines the files to load
+
+Options
+-------
+
+* -h, --help: show a help screen
+* --version: show program's version number and exit
 """
 # ===========================================================================
 __author__     = "Daniel Krajzewicz"
-__copyright__  = "Copyright 2016-2024, Daniel Krajzewicz"
+__copyright__  = "Copyright 2016-2025, Daniel Krajzewicz"
 __credits__    = ["Daniel Krajzewicz"]
 __license__    = "BSD"
 __version__    = "0.18.0"
@@ -33,34 +38,31 @@ __status__     = "Development"
 
 # --- imports ---------------------------------------------------------------
 import sys
-from optparse import OptionParser
-
+import argparse
+import configparser
+from typing import List
 
 
 # --- methods ---------------------------------------------------------------
 # -- main
-def main(arguments=None):
+def main(arguments : List[str] = None) -> int:
     """Initialises an empty memory, loads subsequently the defined files
     to their addresses and writes the memory as occupied by the loaded
     contents.
 
     Args:
         arguments (List[str]): The command line arguments
-
-    Options
-    -------
-
-    * __--file/-f _&lt;FILE&gt;[,&lt;FILE&gt;]*___: the file(s) to load, separated by ‘,’
-    * __--output/-o _&lt;FILE&gt;___: the name of the file to write the memory to
     """
-    optParser = OptionParser(usage="""usage:\n  %prog <MEMORY_DUMP>\n  %prog [options]""", version="filemerge 0.18.0")
-    optParser.add_option("-f", "--file", dest="file", default=None, help="Defines the files to load, separated by a ','")
-    optParser.add_option("-o", "--output", dest="output", default=None, help="Defines the name of the file to write")
-    options, remaining_args = optParser.parse_args(args=arguments)
-    if not options.file:
+    parser = argparse.ArgumentParser(prog="filemerge", description="Joins a set of files into one.",
+        epilog='(c) Daniel Krajzewicz 2016-2025')
+    parser.add_argument('--version', action='version', version='%(prog)s 0.18.0')
+    parser.add_argument("output", metavar="OUTPUT_FILE", default=None, help="Defines the name of the file to write")
+    parser.add_argument("files", metavar="INPUT_FILE", nargs='+', default=None, help="Defines the files to load")
+    args = parser.parse_args(arguments)
+    if not args.file:
         optParser.error("no input file(s) given...")
         sys.exit()
-    files = options.file.split(",")
+    files = args.file.split(",")
     mem = bytearray([0]*65536)
     minAddr = 65536
     maxAddr = 0
@@ -72,7 +74,7 @@ def main(arguments=None):
         minAddr = min(minAddr, addr)
         maxAddr = max(maxAddr, addr+len(vals)-2)
         mem = mem[:addr] + bytearray(vals[2:]) + mem[addr+len(vals)-2:]
-    fd = open(options.output, "wb")
+    fd = open(args.output, "wb")
     print (minAddr)
     fd.write(bytearray([minAddr%256, int(minAddr/256)]))
     fd.write(mem[minAddr:maxAddr])
@@ -81,4 +83,4 @@ def main(arguments=None):
 
 # -- main check
 if __name__ == "__main__":
-    main(sys.argv)
+    sys.exit(main(sys.argv[1:])) # pragma: no cover