+++ src/mem2png.py
@@ -1,21 +1,26 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-from __future__ import print_function
-# ===========================================================================
 """c64tools - mem2png - A c64 memory dump visualiser that can export the dump to an image.
 
-Either one argument - the file to load - must be supported or the following options
+The file to load must be supported. Either --show or --output must be set.
 
---file/-f: the memory dump to load
---output/-o: the image file to write
---width/-w: the width of the window / image in characters
+Arguments
+---------
 
-(c) Daniel Krajzewicz 2016-2024
-daniel@krajzewicz.de
+* MEMORY_FILE: the memory dump file to load
+
+Options
+-------
+
+* --show/-s OUTPUT_FILE: show the memory dump
+* --output/-o OUTPUT_FILE: the name/path of the output file
+* --width/-w INT: the width of the window in chars (the default is 128)
+* -h, --help: show a help screen
+* --version: show program's version number and exit
 """
 # ===========================================================================
 __author__     = "Daniel Krajzewicz"
-__copyright__  = "Copyright 2016-2024, Daniel Krajzewicz"
+__copyright__  = "Copyright 2016-2025, Daniel Krajzewicz"
 __credits__    = ["Daniel Krajzewicz"]
 __license__    = "BSD"
 __version__    = "0.18.0"
@@ -33,23 +38,19 @@ __status__     = "Development"
 import os
 import sys
 sys.path.append(os.path.dirname(os.path.realpath(__file__)))
+import argparse
+import configparser
 import c64tools
+from typing import List
 
 
 # --- methods ---------------------------------------------------------------
 # -- main
-def main(arguments=None):
+def main(arguments : List[str] = None) -> int:
     """Loads the memory dump, saves its display and shows it.
 
     Args:
         arguments (List[str]): The command line arguments
-
-    Options
-    -------
-
-    * __--file/-f _&lt;MEMORY_DUMP_TO_LOAD&gt;___: the memory dump to load
-    * __--output/-o _&lt;IMAGE_TO_SAVE&gt;___: the file name to save the memory image to
-    * __--width/-w _&lt;DISPLAY_WIDTH_IN_CHARACTERS&gt;___: the width of the display in characters (the default is 128)
     """
     import pygame
     # check arguments
@@ -65,23 +66,21 @@ def main(arguments=None):
         ifile = arguments[1]
         show = True
     else:
-        from optparse import OptionParser
-        optParser = OptionParser(usage="""usage:\n  %prog <MEMORY_DUMP>\n  %prog [options]""", version="mem2png 0.18.0")
-        optParser.add_option("-f", "--file", dest="file", default=None, help="Defines the memory dump file to load")
-        #optParser.add_option("-s", "--show", dest="show", action="store_true", default=False, help="Show the dump after loading")
-        optParser.add_option("-o", "--output", dest="store", default=None, help="Defines the name/path of the output file")
-        optParser.add_option("-w", "--width", dest="width", type="int", default=128, help="Defines the width of the window in chars")
-        options, remaining_args = optParser.parse_args(args=arguments)
-        ifile = options.file
-        ofile = options.store
-        #show = options.show
-        xs = options.width
-        #if not ofile and not show:
-        #  print "You should either want to display (--show) or to save (--output) the file"
-        #  return
-        if ifile is None:
-            optParser.error("No input file(s) given...")
-            sys.exit(2)
+        parser = argparse.ArgumentParser(prog="mem2png", 
+            description="A c64 memory dump visualiser that can export the dump to an image.",
+            epilog='(c) Daniel Krajzewicz 2016-2025')
+        parser.add_argument('--version', action='version', version='%(prog)s 0.18.0')
+        parser.add_argument("input", metavar="MEMORY_FILE", default=None, help="the memory dump file to load")
+        parser.add_argument("-s", "--show", dest="show", action="store_true", default=False, help="show the memory dump")
+        parser.add_argument("-o", "--output", dest="store", default=None, help="the name/path of the output file")
+        parser.add_argument("-w", "--width", dest="width", type=int, default=128, help="the width of the window in chars (the default is 128)")
+        args = parser.parse_args(arguments)
+        ifile = args.input
+        ofile = args.store
+        show = args.show
+        xs = args.width
+        if not ofile and not show:
+            raise argparse.ArgumentTypeError("Either --show or the output file (--output <FILE>) must be set.")
     # build window and load and draw dump
     INFO = 40
     ys = int(65536/xs)
@@ -103,21 +102,22 @@ def main(arguments=None):
         pygame.image.save(s, ofile)
 
     # run the display loop
-    while (w.show):
-        pygame.display.update()
-        mx,my = pygame.mouse.get_pos()
-        char = (int(mx / 8) + (int(my/8)*int(xs/8))) * 8
-        pygame.gfxdraw.filled_polygon(w.screen, [[0,ys+1], [xs,ys+1], [xs,ys+INFO], [0,ys+INFO], [0,ys+1]], (0,0,0))
-        textsurface = myfont.render('x=%s, y=%s' % (mx,my), False, (255,255,255))
-        w.screen.blit(textsurface, (0, ys))
-        textsurface = myfont.render('char address: %s (%x)' % (char, char), False, (255,255,255))
-        w.screen.blit(textsurface, (0, ys+16))
-        pygame.display.flip()
-        w.run()
+    if args.show:
+        while (w.show):
+            pygame.display.update()
+            mx,my = pygame.mouse.get_pos()
+            char = (int(mx / 8) + (int(my/8)*int(xs/8))) * 8
+            pygame.gfxdraw.filled_polygon(w.screen, [[0,ys+1], [xs,ys+1], [xs,ys+INFO], [0,ys+INFO], [0,ys+1]], (0,0,0))
+            textsurface = myfont.render('x=%s, y=%s' % (mx,my), False, (255,255,255))
+            w.screen.blit(textsurface, (0, ys))
+            textsurface = myfont.render('char address: %s (%x)' % (char, char), False, (255,255,255))
+            w.screen.blit(textsurface, (0, ys+16))
+            pygame.display.flip()
+            w.run()
     pygame.quit()
 
 
 # -- main check
 if __name__ == "__main__":
-    main(sys.argv)
+    sys.exit(main(sys.argv[1:])) # pragma: no cover